ml_service/requirements.txt
fastapi
uvicorn
pydantic
requests
numpy
pandas
scikit-learn
scikit-surprise
pydantic-settings
psycopg2-binary
